chore: modularize Button componnent
